# MNT: Refactor flight prints module